requirements.txt
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
transformers==4.29.2
timm==0.9.2
Flask==2.3.2